Fixed build on MacOS with clang in C++11 mode.
avfvideowidget.mm:120:9: error: non-constant-expression cannot be narrowed from type 'int' to 'GLfloat' (aka 'float') in initializer list [-Wc++11-narrowing]
        x1, y1, zValue,
                ^~

+++ src/plugins/avfoundation/mediaplayer/avfvideowidget.mm
@@ -103,11 +103,11 @@ void AVFVideoWidget::paintGL()
         return;
 
     QRect targetRect = displayRect();
-    int x1 = targetRect.left();
-    int x2 = targetRect.right();
-    int y1 = targetRect.bottom();
-    int y2 = targetRect.top();
-    int zValue = 0;
+    GLfloat x1 = targetRect.left();
+    GLfloat x2 = targetRect.right();
+    GLfloat y1 = targetRect.bottom();
+    GLfloat y2 = targetRect.top();
+    GLfloat zValue = 0;
 
     const GLfloat textureCoordinates[] = {
         0, 0,
